+++ src/shared/components/case-viewer/index.ts
@@ -6,3 +6,4 @@ export * from './case-viewer.component';
 export * from './case-viewer.module';
 export * from './viewer.routing';
 export * from './case-challenged-access-request';
+export * from './case-challenged-access-success';