+++ RELEASE-NOTES.md
@@ -1,4 +1,6 @@
 ## RELEASE NOTES
+### Version 4.3.1-EUI-4202-challenged-access-request-screen-functionality
+**EUI-4202** Challenged Access Request screen functionality - happy path
 
 ### Version 4.5.0-rc.4
 **EUI-4415** Label not rendering